+++ settings/shared/enriched/reload_user_profile.sql
@@ -4,26 +4,36 @@
   })
 }}
 
-with users_activity as (
+with users_registration as (
   select  product_id, user_id, install_time, install_id
         , agency, media_source, campaign_id, country_code, platform, os_version, device_model
-        , registration_time, first_login_time, last_login_time, session_count, level
+        , registration_time
   from (
     select  product_id, user_id, install_time, install_id
           , agency, media_source, campaign_id, country_code, platform, os_version, device_model
           , registration_time, row_number() over (partition by product_id, user_id order by registration_time) as rn
-          , min(first_login_time) over(partition by product_id, user_id) as first_login_time
-          , max(last_login_time) over(partition by product_id, user_id) as last_login_time
-          , sum(session_count) over(partition by product_id, user_id) as session_count
-          , max(level) over(partition by product_id, user_id) as level
     from {{
       source({
-        'name': 'users_activity',
-        'location': '{{lakehouse.location}}/{{product_id}}/enriched/user/activity',
+        'name': 'users_registration',
+        'location': '{{lakehouse.location}}/{{product_id}}/enriched/user/registration',
       })
     }}
   )
   where rn = 1
+)
+, users_activity as (
+  select  product_id, user_id
+        , min(first_login_time) as first_login_time
+        , max(last_login_time) as last_login_time
+        , sum(session_count) as session_count
+        , max(level) as level
+  from {{
+    source({
+      'name': 'users_activity',
+      'location': '{{lakehouse.location}}/{{product_id}}/enriched/user/activity',
+    })
+  }}
+  group by product_id, user_id
 )
 , users_purchase as (
   select  product_id, user_id
@@ -44,5 +54,6 @@ select  product_id, user_id, install_time, install_id
       , agency, media_source, campaign_id, country_code, platform, os_version, device_model
       , registration_time, first_login_time, last_login_time, session_count, level
       , first_purchase_time, last_purchase_time, number_of_purchases, revenue_usd
-from users_activity
+from users_registration
+full join users_activity using (product_id, user_id)
 full join users_purchase using (product_id, user_id)